fixes for new hosting

+++ deploy.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
-SERV=xonix@rt.mipt.ru
-TOMCAT=/usr/local/apache-tomcat-7.0/
+#SERV=xonix@rt.mipt.ru
+#TOMCAT=/usr/local/apache-tomcat-7.0/
+SERV=xonix@37.187.123.144
+TOMCAT=/home/xonix/apache-tomcat/
 WAR=$(ls -v zlo_web/target/*.war | tail -n 1)
 BKP=$(date +%y%m%d%H%M)
 
@@ -15,7 +17,11 @@ $TOMCAT/bin/shutdown.sh; sleep 10; killall -9 java
 echo 'Clean/Backup...'
 
 rm -rf $WA/ROOT
-mv $WA/ROOT.war $WA/ROOT.war_$BKP
+
+if [ -f $WA/ROOT.war ]
+then
+    mv $WA/ROOT.war $WA/ROOT.war_$BKP
+fi
 
 rm -rf $TOMCAT/work/*
 "
@@ -24,9 +30,6 @@ scp $WAR $SERV:$WA/ROOT.war
 
 ssh $SERV "
 echo 'Starting...'
-
-CATALINA_OPTS='-server -Xmx300m -XX:MaxPermSize=256m -XX:PermSize=128m'
-export CATALINA_OPTS
 
 $TOMCAT/bin/startup.sh;
